Generate multiple documentation sets
We'd like both public and internal sets of documentation. This allows us
to generate those from the same Swagger template file.

We've extended the Swagger "language" to include the notion of
documentation sets. Each docset represents a target that we want to
build.

At the top level we define an array of `documentationSets`:

    "documentationSets": [
      "public",
      "internal"
    ]

Each endpoint (or "path") has an `onlyVisibleInDocumentationSets` key
that defines which documentation that endpoint appears in. So if we had
a `/secret_stuff` endpoint that shouldn't be public, we'd add:

    "onlyVisibleInDocumentationSets": [
      "internal"
    ]

Each endpoint defaults to being visible in all documentation sets.

This is implemented as a Ruby script and can be run through a Makefile.
The Makefile also include a `validate` task, which runs `swagger
validate` on the generated documentation to ensure that it's
well-formed.